+++ tests/src/Kernel/NbspCleanerFilterTest.php
@@ -51,11 +51,6 @@ class NbspCleanerFilterTest extends KernelTestBase {
     $result = $filter->process($input, 'und');
     $this->assertInstanceOf(FilterProcessResult::class, $result);
 
-    // Since Drupal 10.2.0 use filter system HTML5.
-    if (version_compare(\Drupal::VERSION, '10.2.0', '>')) {
-      $expected = str_replace(' ', '&nbsp;', $expected);
-    }
-
     $this->assertEquals($expected, $result->getProcessedText());
   }
 
@@ -68,23 +63,23 @@ class NbspCleanerFilterTest extends KernelTestBase {
       ['<p>Maecenas cursus posuere</p>', '<p>Maecenas cursus posuere</p>'],
       [
         '<p>Maecenas<nbsp>&nbsp;</nbsp>cursus posuere</p>',
-        '<p>Maecenas cursus posuere</p>',
+        '<p>Maecenas&nbsp;cursus posuere</p>',
       ],
       [
         '<p>Maecenas <a href="https://www.google.ch">lorem<nbsp>&nbsp;</nbsp>ipsum</a><nbsp>&nbsp;</nbsp>cursus<nbsp>&nbsp;</nbsp>posuere</p>',
-        '<p>Maecenas <a href="https://www.google.ch">lorem ipsum</a> cursus posuere</p>',
+        '<p>Maecenas <a href="https://www.google.ch">lorem&nbsp;ipsum</a>&nbsp;cursus&nbsp;posuere</p>',
       ],
       [
         '<p>Maecenas<nbsp>&nbsp;</nbsp>cursus<nbsp>&nbsp;</nbsp>posuere</p>',
-        '<p>Maecenas cursus posuere</p>',
+        '<p>Maecenas&nbsp;cursus&nbsp;posuere</p>',
       ],
       [
         '<p>Maecenas<div class="nbsp">&nbsp;</div>cursus posuere</p>',
-        '<p>Maecenas</p><div class="nbsp"> </div>cursus posuere',
+        '<p>Maecenas</p><div class="nbsp">&nbsp;</div>cursus posuere',
       ],
       [
         '<p>Maecenas<span>&nbsp;</span>cursus posuere</p>',
-        '<p>Maecenas<span> </span>cursus posuere</p>',
+        '<p>Maecenas<span>&nbsp;</span>cursus posuere</p>',
       ],
     ];
   }